smoll fix

+++ src/TranspositionTable.cpp
@@ -59,7 +59,7 @@ void TranspositionTable::clear()
 	std::memset(transpositionTable, 0, tableSize * sizeof(TTEntry));
 }
 
-void TranspositionTable::save(uint64_t key, Move move, double evaluation, int depth, int alpha, int beta, Flags flags)
+void TranspositionTable::save(uint64_t key, Move move, int evaluation, int depth, int alpha, int beta, Flags flags)
 {
 	TTEntry newEntry = {key, move, evaluation, depth, alpha, beta, flags};
 	transpositionTable[key % tableSize] = newEntry;
